+++ src/handlers/lobby.handler.js
@@ -7,6 +7,8 @@ import Result from './result.js';
 import { addUser, getUserById } from '../session/user.session.js';
 import { getRedis } from '../redis.js';
 import { v4 as uuidv4 } from 'uuid';
+import { createLobby } from '../db/Lobby/createLobby.js';
+import { joinRoom } from '../db/Lobby/joinRoom.js';
 
 // 환경 변수에서 설정 불러오기
 const { PacketType } = configs;
@@ -34,40 +36,16 @@ string gameUrl = 2;
 */
 export const createLobbyHandler = async ({ socket, payload }) => {
   const { isPrivate } = payload;
-  const redis = await getRedis();
-
-  const player_count = 0;
-  const max_player = 4;
-
-  const GameSession = {
-    game_id: GAME_ID,
-    create_at: Date.now(),
-    update_at: Date.now(),
-    isPrivate: isPrivate,
-    state: LobbyState,
-    max_player: max_player,
-    player_count: player_count,
-  };
-
-  try {
-    const transaction = redis.multi();
-
-    transaction.hset(`GameSession:${GAME_ID}`, GameSession);
-    transaction.rpush('GameSessions', GAME_ID);
-    if (!isPrivate) {
-      transaction.rpush('LobbyQueue', GAME_ID);
-    }
-
-    await transaction.exec();
-  } catch (error){
-    logger.error(`Failed to Create Lobby : ${error.message}`);
-  }
-/*
+  const { gameCode, gameUrl } = createLobby(isPrivate);
+  /*
+  dev에서 내걸로 가져와서 문제없는지 구동해보기
+  */
+  /*
 // 방 생성 결과
 message S2CCreateRoomRes {
 string gameCode = 1; // 방 코드
 string gameUrl = 2;
 }
 */
-  return new Result({ gameCode : GAME_ID, gameUrl : "이거 주소임" }, PacketType.CREATE_ROOM_RESPONSE);
+  return new Result({ gameCode, gameUrl }, PacketType.CREATE_ROOM_RESPONSE);
 };
